Clean up installs_item_paths
It just caught my eye that the 'installs_item_paths' in the MunkiInstallsItemsCreator has some '/usr/local/sal/yaml' paths, which were needed for sal-scripts 2, but no longer for sal-scripts 3, because these paths have been moved to puppet_checkin_module (which I can't get to work atm, but I'll make a separate issue for that later). There was also a typo (/usr/local/munki/postflight.d was listed twice). This patch cleans up the installs_item_paths and fixes the typo, and should not affect the functionality of this recipe.